+++ internal/provider/data_source_coralogix_events2metric_test.go
@@ -27,6 +27,7 @@ func TestAccCoralogixDataSourceEvents2Metric_basic(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckEvents2MetricDestroy,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCoralogixResourceLogs2Metric(logsToMetric, "") +